+++ HiveWays.Core/HiveWays.Domain/Entities/DataPointEntity.cs
@@ -7,13 +7,13 @@ public class DataPointEntity : ITableEntity
 {
     public string PartitionKey { get; set; }
     public string RowKey { get; set; }
-    public decimal Longitude { get; set; }
-    public decimal Latitude { get; set; }
-    public decimal SpeedMps { get; set; }
-    public decimal SpeedKmph { get; set; }
-    public decimal Heading { get; set; }
-    public decimal AccelerationMps { get; set; }
-    public decimal AccelerationKmph { get; set; }
+    public double Longitude { get; set; }
+    public double Latitude { get; set; }
+    public double SpeedMps { get; set; }
+    public double SpeedKmph { get; set; }
+    public double Heading { get; set; }
+    public double AccelerationMps { get; set; }
+    public double AccelerationKmph { get; set; }
 
     public DateTimeOffset? Timestamp { get; set; }
     public ETag ETag { get; set; }
