+++ HiveWays.Core/HiveWays.Domain/Models/DataPoint.cs
@@ -11,17 +11,17 @@ public class DataPoint : IIdentifiable
     public int Id { get; set; }
 
     [Name("x[m]")]
-    public decimal X { get; set; }
+    public double X { get; set; }
 
     [Name("y[m]")]
-    public decimal Y { get; set; }
+    public double Y { get; set; }
 
     [Name("speed[m/s]")]
-    public decimal Speed { get; set; }
+    public double Speed { get; set; }
 
     [Name("heading")]
-    public decimal Heading { get; set; }
+    public double Heading { get; set; }
 
     [Name("acc[m/s^2]")]
-    public decimal Acceleration { get; set; }
+    public double Acceleration { get; set; }
 }
