+++ HiveWays.Core/HiveWays.Domain/Models/VehicleStats.cs
@@ -4,9 +4,9 @@ public class VehicleStats : IIdentifiable
 {
     public int Id { get; set; }
     public DateTime Timestamp { get; set; }
-    public decimal Longitude { get; set; }
-    public decimal Latitude { get; set; }
-    public decimal SpeedKmph { get; set; }
-    public decimal Heading { get; set; }
-    public decimal AccelerationKmph { get; set; }
+    public double Longitude { get; set; }
+    public double Latitude { get; set; }
+    public double SpeedKmph { get; set; }
+    public double Heading { get; set; }
+    public double AccelerationKmph { get; set; }
 }
