+++ HiveWays.FleetIntegration/Business/ClusterConfiguration.cs
@@ -2,9 +2,9 @@
 
 public class ClusterConfiguration
 {
-    public decimal ClusterRadius { get; set; }
+    public double ClusterRadius { get; set; }
     public int VehicleLength { get; set; }
     public int VehicleWidth { get; set; }
     public int VehicleDistance { get; set; }
-    public decimal OrientationLimit { get; set; }
+    public double OrientationLimit { get; set; }
 }
